+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "terminal.integrated.env.osx": {
     "PATH": "/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/homebrew/bin"
-  }
+  },
+  "search.followSymlinks": false
 }
