+++ CMakeLists.txt
@@ -24,7 +24,7 @@ set(CMAKE_CXX_STANDARD 17)
 
 add_subdirectory(third_party/lineairdb)
 set(LINEAIRDB_PLUGIN_DYNAMIC "ha_lineairdb")
-set(LINEAIRDB_SOURCES ha_lineairdb.cc)
+set(LINEAIRDB_SOURCES ha_lineairdb.cc ha_lineairdb_handler_miscs.cc)
 add_definitions(-DMYSQL_SERVER)
 
 if(WITH_LINEAIRDB_STORAGE_ENGINE AND NOT WITHOUT_LINEAIRDB_STORAGE_ENGINE)
